Hiremi/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
djangorestframework_simplejwt @ git+https://github.com/SimpleJWT/django-rest-framework-simplejwt.git@3b71b89779043acd865f13547d88b8355e4ef525
dnspython==2.6.1
idna==3.7
paytm-pg==1.1.1
paytmchecksum==1.7.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
validate_email==1.3
